server/requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
fastapi==0.75.1
h11==0.13.0
httptools==0.4.0
idna==3.3
pydantic==1.9.0
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
torch==1.8.0
torchvision==0.9.0
pip~=22.0.4
wheel~=0.36.2
setuptools~=57.0.0
einops==0.4.1
ftfy~=6.1.1
regex~=2022.3.15
tqdm~=4.63.1
requests~=2.27.1
Pillow~=9.1.0
numpy~=1.22.3
python-multipart==0.0.5
loguru==0.6.0
anyio==3.5.0
asgiref==3.5.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
einops==0.4.1
fastapi==0.75.1
firebase-admin==5.2.0
ftfy==6.1.1
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
loguru==0.6.0
msgpack==1.0.3
numpy==1.22.3
Pillow==9.1.0
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.7
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
rsa==4.8
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
torch==1.8.0
torchvision==0.9.0
tqdm==4.63.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.2
anyio==3.5.0
asgiref==3.5.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
dnspython==2.2.1
einops==0.4.1
email-validator==1.1.3
fastapi==0.75.1
firebase-admin==5.2.0
ftfy==6.1.1
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
loguru==0.6.0
msgpack==1.0.3
numpy==1.22.3
Pillow==9.1.0
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.0
pyparsing==3.0.7
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
rsa==4.8
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
torch==1.8.0
torchvision==0.9.0
tqdm==4.63.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.2
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
dnspython==2.2.1
einops==0.4.1
email-validator==1.1.3
fastapi==0.75.1
firebase-admin==5.2.0
ftfy==6.1.1
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
loguru==0.6.0
msgpack==1.0.3
numpy==1.22.3
passlib==1.7.4
Pillow==9.1.0
proto-plus==1.20.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
rsa==4.8
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
torch==1.8.0
torchvision==0.9.0
tqdm==4.63.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.2
